tema-codare-jpeg/4.py
import cv2 
import numpy as np
from jpeg_color import JPEGColor

def extract_frames(video_path):
    cap = cv2.VideoCapture(video_path)
    frames = []
    while True:
        ret, frame = cap.read()
        if not ret:
            break
        frame_rgb = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
        frames.append(frame_rgb)
    
    cap.release()
    return frames


def build_video(frames, output_path, fps=30):
    h, w, _ = frames[0].shape
    fourcc = cv2.VideoWriter_fourcc(*"mp4v")
    writer = cv2.VideoWriter(output_path, fourcc, fps, (w, h))

    for frame_rgb in frames:
        frame_bgr = cv2.cvtColor(frame_rgb, cv2.COLOR_RGB2BGR)
        writer.write(frame_bgr)

    writer.release()
    return output_path

if __name__ == '__main__':
    video_path = "./res/video.mp4"
    frames = extract_frames(video_path)

    print("Encoding every frame...")
    encoded_frames = []
    for frame in frames:
        jpeg = JPEGColor(frame)
        encoded = jpeg.encode()
        encoded_frames.append(encoded)
    
    print("Building encoded video...")
    
    output_path = "./output/output.mp4"
    build_video(encoded_frames, output_path)
    
    print(f"Saved encoded video as {output_path}")